+++ bank-feeds/RELEASES.md
@@ -99,3 +99,13 @@ Based on:
 - [csharp v2.2.0] bank-feeds
 ### Releases
 - [NuGet v2.2.0] https://www.nuget.org/packages/Codat.BankFeeds/2.2.0 - bank-feeds
+
+## 2023-10-23 01:27:15
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.104.0 (2.169.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v2.3.0] bank-feeds
+### Releases
+- [NuGet v2.3.0] https://www.nuget.org/packages/Codat.BankFeeds/2.3.0 - bank-feeds